added arbitrary args and some refactoring
added kwargs to add_append_only_table
